requirements.txt
autopep8==1.7.0
bidict==0.22.0
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
Flask-SocketIO==5.3.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
pycodestyle==2.9.1
python-engineio==4.3.4
python-socketio==5.7.1
toml==0.10.2
Werkzeug==2.2.2
zipp==3.8.1
